requirements.txt
requests==2.32.*
python-dotenv==1.2.*
terminaltables==3.1.*
environs==14.6.0